refactor: move rpc to a lib
