+++ src/utils/workers/index.ts
@@ -1,0 +1,1 @@
+export type CleanupCallback = () => void;
